[WIP] Starting DXR implementation

+++ GadgetEngine/Graphics/DX12/DX12_Helpers.cpp
@@ -86,7 +86,7 @@ void DX12_Helpers::DX12_ResourceBarriers::ApplyAllBarriers(ID3D12_GraphicsComman
 //-------------------- Buffers -----------------------------------------------------------------------//
 //----------------------------------------------------------------------------------------------------//
 
-ID3D12Resource* DX12_Helpers::CreateBuffer(ID3D12_Device* device_, const void* data_, uint32_t bufferSize_, bool isCpuAccessible_, D3D12_RESOURCE_STATES state_, D3D12_RESOURCE_FLAGS flags_, ID3D12Heap* heap_, uint64_t heapOffset_){
+ID3D12Resource* DX12_Helpers::CreateBuffer(ID3D12_Device* device_, const void* data_, uint64_t bufferSize_, bool isCpuAccessible_, D3D12_RESOURCE_STATES state_, D3D12_RESOURCE_FLAGS flags_, ID3D12Heap* heap_, uint64_t heapOffset_){
 	GADGET_BASIC_ASSERT(device_ != nullptr);
 	GADGET_BASIC_ASSERT(bufferSize_ > 0);
